pyproject.toml
[project]
name = "csf-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "celery>=5.3.0",
    "cryptography>=46.0.3",
    "email-validator>=2.3.0",
    "fastapi>=0.121.3",
    "google-auth>=2.43.0",
    "greenlet>=3.2.4",
    "jinja2>=3.1.0",
    "mailchimp-marketing>=3.0.80",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.1.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "redis>=5.0.0",
    "reportlab>=4.4.5",
    "requests>=2.32.5",
    "sendgrid>=6.11.0",
    "sqlalchemy[asyncio]>=2.0.44",
    "stripe>=14.0.1",
    "twilio>=9.8.7",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "httpx>=0.28.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
